+++ ReleaseNotes.md
@@ -1,7 +1,13 @@
-## 0.3 - In Progress
+## 0.4 - In Progress
   * TBA
 
-## 0.2 - <small>2014/07/21
+## 0.3 - <small>2014/08/06</small> 
+  * Added `IManageRoles` unit tests.
+  * Added LightSpeed's `UserAuth` unit tests.
+  * Implemented `IClearable` in LightSpeedUserAuthRepository.
+  * Fixed various roles management-related method persistence issues.
+
+## 0.2 - <small>2014/07/21</small>
   * Added write compatibility unit tests.
   * Implemented `IManageRoles` in LightSpeedUserAuthRepository.
   * Fixed `CreateUserAuth` and `CreateOrMergeUserAuth` persistence methods.